requirements.txt
pandas
sqlalchemy
pymysql
python-dotenv
numpy
matplotlib.pyplot